Now services throw HTTP errors and controllers just call them. Also DTOs are more thoroughly decorated.

+++ packages/server/src/services/carpool.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Carpool } from "../entities";
 import { Repository } from "typeorm";
@@ -18,16 +18,9 @@ export class CarpoolService {
         const { destination, eventId } = createCarpoolDto;
         let carpoolName = createCarpoolDto.carpoolName;
 
-        if (!destination) {
-            throw new Error("Carpools require a destination");
-        }
-        if (!eventId) {
-            throw new Error("An eventId must be specified when creating a Carpool");
-        }
-
         const event = await this._eventService.get(eventId, true);
         if (!event) {
-            throw new Error("No event could be found with the provided eventId");
+            throw new NotFoundException("No event could be found with the provided eventId");
         }
         if (!carpoolName) {
             carpoolName = `${event.name} carpool`;
@@ -43,19 +36,25 @@ export class CarpoolService {
 
         const carpool = new Carpool();
         carpool.name = createCarpoolDto.carpoolName;
-        carpool.destination = createCarpoolDto.destination;
+        carpool.destination = destination;
         carpool.eventId = createCarpoolDto.eventId;
 
         return await this._carpoolRepository.save(carpool);
     }
 
     public async get(id: string): Promise<Carpool> {
-        let carpool = await this._carpoolRepository.findOneOrFail(id);
+        let carpool = await this._carpoolRepository.findOne(id);
+        if (!carpool) {
+            throw new NotFoundException("No Carpool was found with the provided ID");
+        }
         return carpool;
     }
 
     public async update(id: string, updateCarpoolDto: UpdateCarpoolDto): Promise<Carpool> {
-        let carpool = await this._carpoolRepository.findOneOrFail(id);
+        let carpool = await this._carpoolRepository.findOne(id);
+        if (!carpool) {
+            throw new NotFoundException("No Carpool was found with the provided ID");
+        }
 
         //TODO: Add an entity mapper for obvious reasons...
         carpool.name = updateCarpoolDto.carpoolName;
@@ -67,7 +66,10 @@ export class CarpoolService {
 
     //TODO: Comment this :)
     public async delete(id: string): Promise<Carpool> {
-        let carpool = await this._carpoolRepository.findOneOrFail(id);
+        let carpool = await this._carpoolRepository.findOne(id);
+        if (!carpool) {
+            throw new NotFoundException("No Carpool was found with the provided ID");
+        }
         return await this._carpoolRepository.remove(carpool);
     }
 }
